+++ mini-micoservices.sln
@@ -31,8 +31,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App.Web.Client", "src\apps\
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Shared", "Shared", "{56DC7805-5C68-4B40-B39C-7DC8C6BC7ABA}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App.Common.Abstractions", "src\shared\App.Common.Abstractions\App.Common.Abstractions.csproj", "{4FD13424-2BA2-1233-41F9-CB918776CB44}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App.Common.Infrastructure", "src\shared\App.Common.Infrastructure\App.Common.Infrastructure.csproj", "{D679E202-5187-2B05-2F11-29990513A7D0}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "App.Common.Domain", "src\shared\App.Common.Model\App.Common.Domain.csproj", "{3976FD95-049C-1962-9227-D2D59F01C86A}"
@@ -63,10 +61,6 @@ Global
 		{B1D3D80B-262E-447A-4D5B-20751EF14B0C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B1D3D80B-262E-447A-4D5B-20751EF14B0C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B1D3D80B-262E-447A-4D5B-20751EF14B0C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4FD13424-2BA2-1233-41F9-CB918776CB44}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4FD13424-2BA2-1233-41F9-CB918776CB44}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4FD13424-2BA2-1233-41F9-CB918776CB44}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4FD13424-2BA2-1233-41F9-CB918776CB44}.Release|Any CPU.Build.0 = Release|Any CPU
 		{D679E202-5187-2B05-2F11-29990513A7D0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{D679E202-5187-2B05-2F11-29990513A7D0}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D679E202-5187-2B05-2F11-29990513A7D0}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -90,7 +84,6 @@ Global
 		{5E6944B9-7CE3-E953-8612-07EE42081E95} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{B1D3D80B-262E-447A-4D5B-20751EF14B0C} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{56DC7805-5C68-4B40-B39C-7DC8C6BC7ABA} = {536C8147-8CE3-4B92-AB66-E9DEB707BD29}
-		{4FD13424-2BA2-1233-41F9-CB918776CB44} = {56DC7805-5C68-4B40-B39C-7DC8C6BC7ABA}
 		{D679E202-5187-2B05-2F11-29990513A7D0} = {56DC7805-5C68-4B40-B39C-7DC8C6BC7ABA}
 		{3976FD95-049C-1962-9227-D2D59F01C86A} = {56DC7805-5C68-4B40-B39C-7DC8C6BC7ABA}
 	EndGlobalSection
