+++ src/apps/App.Api.Todo/Features/Tags/Services/ITagService.cs
@@ -1,4 +1,4 @@
-﻿using App.Common.Domain.Dtos;
+﻿using App.Common.Domain.Dtos.Todo;
 
 namespace App.Api.Todo.Features.Tags.Services
 {
