+++ src/apps/App.Api.Todo/Features/Todotask/Data/ITodoTaskRepository.cs
@@ -1,4 +1,4 @@
-﻿using App.Common.Domain.Dtos;
+﻿using App.Common.Domain.Dtos.Todo;
 using App.Common.Infrastructure.Model;
 using TodoTask = App.Api.Todo.Models.TodoTask;
