+++ src/apps/App.Api.Todo/Features/Todotask/Endpoints/TodoTaksEndpoints.cs
@@ -1,6 +1,5 @@
-﻿using App.Api.Todo.Features.Todotask.Dtos;
-using App.Api.Todo.Features.Todotask.Services;
-using App.Common.Domain.Dtos;
+﻿using App.Api.Todo.Features.Todotask.Services;
+using App.Common.Domain.Dtos.Todo;
 
 namespace App.Api.Todo.Features.Todotask.Endpoints
 {
