+++ src/apps/App.Api.Todo/Features/Todotask/Mapper/ITaskMapper.cs
@@ -1,4 +1,5 @@
 ﻿using App.Api.Todo.Features.Todotask.Dtos;
+using App.Common.Domain.Dtos.Todo;
 using TodoTask = App.Api.Todo.Models.TodoTask;
 
 namespace App.Api.Todo.Features.Todotask.Mapper
