+++ src/apps/App.Api.Todo/Features/Todotask/Services/TodoTaskService.cs
@@ -1,7 +1,7 @@
 ﻿using App.Api.Todo.Features.Todotask.Data;
 using App.Api.Todo.Features.Todotask.Dtos;
 using App.Api.Todo.Features.Todotask.Mapper;
-using App.Common.Domain.Dtos;
+using App.Common.Domain.Dtos.Todo;
 using App.Common.Domain.Pagination;
 using Microsoft.EntityFrameworkCore;
