+++ src/apps/App.Api.Todo/appsettings.json
@@ -7,7 +7,7 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "TodoDb": "Host=localhost;Port=5432;Database=mydatabase;Username=sa;Password=YourStrong!Pass123"
+    "TodoDb": "Host=localhost;Port=5432;Database=mydatabase;Username=postgres;Password=YourStrong!Pass123"
   },
   "Entra": {
     "TenantId": "",
