+++ src/apps/App.Web.Client/appsettings.json
@@ -11,7 +11,7 @@
     "Authority": "https://login.microsoftonline.com/7d49edce-eda8-4d56-8fe8-04390a0eb982/v2.0",
     "TenantId": "7d49edce-eda8-4d56-8fe8-04390a0eb982",
     "ClientId": "beedb05b-78f1-4917-a0c8-271cf7188ac3",
-    "ClientSecret": "jMG8Q~RHAZ5tHOFcL0W4zxESKHTshB1k28BejcRK",
+    "ClientSecret": "",
     "CallbackPath": "/signin-oidc"
   }
 }
