+++ src/shared/App.Common.Model/Dtos/Todo/TaskDto.cs
@@ -1,4 +1,4 @@
-﻿using App.Common.Domain.Dtos;
+﻿using App.Common.Domain.Dtos.Todo;
 
 namespace App.Api.Todo.Features.Todotask.Dtos
 {
