+++ tests/Test.App.Api.Todo.IntegrationTest/FeatureTests/TodotaskTests/EndpointTests/CreateTodoTaskTests.cs
@@ -1,5 +1,6 @@
 ﻿using App.Api.Todo.Features.Todotask.Dtos;
 using App.Api.Todo.Models;
+using App.Common.Domain.Dtos.Todo;
 using Shouldly;
 using System.Net.Http.Json;
 using Test.App.Api.Todo.IntegrationTest.Fixture;
